ci: run stdlib tests for other architectures
Currently this runs the tests for amd64, 386, arm, arm64, and wasi. I
tried adding wasm and a baremetal target (such as riscv-qemu) but they
don't pass the tests yet and should be fixed in the future. At least
this change should make sure that existing tests continue to pass on a
variety of architectures.